perf: Update GTF merging and filtering pipeline
In this commit, I update the GTF merging pipeline to simplify the stringtie - gffcompare - filtering steps.
I updated the code so that gffcompare is used to merge all the outputs of stringtie, without running stringtie --merge in between. I then also updated the GTF filtering and annotating script to make use of the outputs of the gffcompare merging step.